+++ app/server.jsx
@@ -1,12 +1,12 @@
+import axios from 'axios';
 import React from 'react';
 import { renderToString } from 'react-dom/server';
-import { RouterContext, match, createMemoryHistory } from 'react-router'
-import axios from 'axios';
+import { createMemoryHistory, match, RouterContext } from 'react-router';
 import { Provider } from 'react-redux';
-import createRoutes from 'routes.jsx';
+import createRoutes from 'routes';
 import configureStore from 'store/configureStore';
+import preRenderMiddleware from 'middlewares/preRenderMiddleware';
 import header from 'components/Meta';
-import { fetchComponentDataBeforeRender } from 'api/fetchComponentDataBeforeRender';
 
 const clientConfig = {
   host: process.env.HOSTNAME || 'localhost',
@@ -22,24 +22,23 @@ axios.defaults.baseURL = `http://${clientConfig.host}:${clientConfig.port}`;
  * and pass it into the Router.run function.
  */
 export default function render(req, res) {
-    const history = createMemoryHistory();
-    const authenticated = req.isAuthenticated();
-    const store = configureStore({
-      user: {
-        authenticated,
-        isWaiting: false,
-        message: '',
-        isLogin: true
-      }
-    }, history);
-
-    const routes = createRoutes(store);
+  const authenticated = req.isAuthenticated();
+  const history = createMemoryHistory();
+  const store = configureStore({
+    user: {
+      authenticated,
+      isWaiting: false,
+      message: '',
+      isLogin: true
+    }
+  }, history);
+  const routes = createRoutes(store);
 
   /*
    * From the react-router docs:
    *
    * This function is to be used for server-side rendering. It matches a set of routes to
-   * a location, without rendering, and calls a callback(error, redirectLocation, renderProps)
+   * a location, without rendering, and calls a callback(err, redirect, props)
    * when it's done.
    *
    * The function will create a `history` for you, passing additional `options` to create it.
@@ -49,27 +48,26 @@ export default function render(req, res) {
    * however you like.
    *
    * The three arguments to the callback function you pass to `match` are:
-   * - error: A javascript Error object if an error occured, `undefined` otherwise.
-   * - redirectLocation: A `Location` object if the route is a redirect, `undefined` otherwise
-   * - renderProps: The props you should pass to the routing context if the route matched, `undefined`
-   *                otherwise.
+   * - err:       A javascript Error object if an error occured, `undefined` otherwise.
+   * - redirect:  A `Location` object if the route is a redirect, `undefined` otherwise
+   * - props:     The props you should pass to the routing context if the route matched,
+   *              `undefined` otherwise.
    * If all three parameters are `undefined`, this means that there was no route found matching the
    * given location.
    */
-  match({ routes, location: req.url }, (error, redirectLocation, renderProps) => {
-    if (error) {
-      res.status(500).send(error.message);
-    } else if (redirectLocation) {
-      res.redirect(302, redirectLocation.pathname + redirectLocation.search);
-    } else if (renderProps) {
-      const InitialView = (
-        <Provider store={store}>
-            <RouterContext {...renderProps} />
-        </Provider>
-      );
-
-      // This method waits for all render component promises to resolve before returning to browser
-      fetchComponentDataBeforeRender(store.dispatch, renderProps.components, renderProps.params)
+  match({routes, location: req.url}, (err, redirect, props) => {
+    if (err) {
+      res.status(500).json(err);
+    } else if (redirect) {
+      res.redirect(302, redirect.pathname + redirect.search);
+    } else if (props) {
+      // This method waits for all render component
+      // promises to resolve before returning to browser
+      preRenderMiddleware(
+        store.dispatch,
+        props.components,
+        props.params
+      )
       .then(() => {
         const initialState = store.getState();
         const componentHTML = renderToString(
@@ -94,11 +92,11 @@ export default function render(req, res) {
           </html>
         `);
       })
-      .catch(() => {
-        res.end(renderFullPage('', {}));
+      .catch((err) => {
+        res.status(500).json(err);
       });
     } else {
-      res.status(404).send('Not Found');
+      res.sendStatus(404);
     }
   });
 }