+++ app/store/configureStore.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import { routerMiddleware } from 'react-router-redux';
 import thunk from 'redux-thunk';
 import rootReducer from 'reducers';
-import promiseMiddleware from 'api/promiseMiddleware';
+import promiseMiddleware from 'middlewares/promiseMiddleware';
 import createLogger from 'redux-logger';
 
 /*