+++ webpack/webpack.config.dev-server.js
@@ -13,7 +13,8 @@ var commonLoaders = [
     // Reason why we put this here instead of babelrc
     // https://github.com/gaearon/react-transform-hmr/issues/5#issuecomment-142313637
     query: {
-      "presets": ["es2015", "react", "stage-0"]
+      "presets": ["es2015", "react", "stage-0"],
+      "plugins": ["transform-decorators-legacy"]
     },
     include: path.join(__dirname, '..', 'app'),
     exclude: path.join(__dirname, '..', 'node_modules')
@@ -56,10 +57,8 @@ module.exports = {
       ])
     },
     resolve: {
+      root: [path.join(__dirname, '..', 'app')],
       extensions: ['', '.js', '.jsx', '.css'],
-      modulesDirectories: [
-        "app", "node_modules"
-      ]
     },
     plugins: [
         new webpack.DefinePlugin({